fix: RandomVariable __getattr__ function must raise AttributeError, not NotImplementedError.
Issue #109 was created to solve a problem detected when operate a Tensor and a `RandomVariable`.
Then, when implemented, it did not solve the problem.
The problem was related to a bug in __getattr__ function overload in `RandomVariable`: if an attribute
is not found, it must raise an AttributeError, not a NotImplementedError.
Finally, the tests has been modified to check this new feature, and the sanitize params function has been removed,
as now it is not necessary to transform `RandomVariable` objects to ed2.RandomVariable or Tensors (register function
will transform the object to a Tensor when necessary).

Close #109.