feat: Created contextmanager package to manage data_model and prob_model contexts.
These contexts allow to define correct random variable shapes. The datamodel context
allow to create expanded models using a sample_shape (plate size). The prob_model context
allow to get access dynaically to the expanded vars when builder function is called in a
probmodel object.
Also, custom exceptions are rised in particular cases, which are defined in the exceptions package.