feat: datamodel can be used outside from a probmodel.
Also, datamodel includes an optional parameter size, to set the sample_shape when building it.
In order to clarify the context for creating the graph and other stuff in probmodels, and now
in datamodels as isolated too, the prob_model builder has been renamed to randvar_registry,
where parameters, variables and dependencies between these elements are registered.

Closes #114, #118, #119, #124.