feat: Code restructured and adapted to inference methods different from VI and SVI.
Previous _last_expanded_{vars|params} are now stored in the inference object. Also,
the loss metric does not expand the models, it is done in the run method from the
inference object. The loss metric receives a the p and q expanded vars and returns
the loss tensor.
Finally, prior, posterior and posterior_predictive functions has been implemented
but raises a NotImplementedError.

Closes #157.