feat: Simplified version of algorithm to sanitize input arguments for Random Variables.
The previous version used tf.cast, tf.broadcast_to and tf.stack to infer the shape of the
arguments and compute the mcm for the shape, and broadcast every argument for the Random
Variable to be in that shape. This behaviour caused the issue described in #142, and cannot be
solved unless every argument for every distribution is treated individually, which right now
is not under the main objectives.
Now, the sanitize function just uses the tf.convert_to_tensor for every Random Variable, from
edward2 or inferpy, even if it is in a list or nested list, which allows to use them as
arguments for other Random Variables.
On the other side, the broadcast and casting now relies on the edward2 implementation, and
therefore the batch_shape argument has been removed.

Closes #142