fix: force specific tfp version. closes #193
In addition of module location modifications, the way that tensors are represented
in the graph changes between minor versions. As we analyze the tf.Graph to detect
dependencies between Random Variables, we need a stable way to create the tf.Graph.

In tfp==0.8.0, tf.Cond is represented in a different way, from the original tensor to a copy
of it (called `tensorname`_1), and the the same structire (cond, broadcasts, etc.). As this
is an experimental way to detect dependencies and we should use a more robust to detect them,
we decided to fix the tfp version to avoid other bugs happening from an increase of this version.

The tfp bumping dependency process should take special care of the tf.Graph, and must be tested
before rising the tfp version. For now, fixed to 0.7.0.